Update curried-form.md
Mostly minor stuff, with one substantive exception: I removed the phrase "the latter" in the eighth text paragraph. That's because "the latter" could easily be misread to refer to the parenthetical mention of the "curry and uncurry functions in the Prelude," which would have the opposite meaning to what the sentence's core is about. Please be sure I have interpreted your intention correctly.