Add a small script to "repermission" readers
Because of a bug in our jsbox integration, GumRoad purchasers are added
as collaborators with Push access. There is an issue open on the jsbox
project[1], but it's received no feedback to date. Up until now I've
been manually going into the GitHub UI and changing the permissions on
collaborators to Pull. This script automates that.

The obvious solution is to fix the integration, but I took this as a
chance to play with the stack interpreter feature for writing one-off
scripts like this (and I'd consider it a huge success). I'm happy to
abandon this if we can get the integration working properly. Also, now
that I've automated this, I'm sure no will ever buy the book again.

[1]: https://github.com/masonforest/jsbox/issues/1